refactor: 레거시 컨트롤러를 어노테이션 기반으로 변경
- Legacy MVC 제거
- DispatcherServlet를 mvc 모듈로 이동
