+++ mvc/src/main/java/com/interface21/webmvc/servlet/DispatcherServletInitializer.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet;
 
 import com.interface21.web.WebApplicationInitializer;
 import jakarta.servlet.ServletContext;
