--- mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/ControllerHandlerMapping.java
@@ -1,34 +1,0 @@
-package com.interface21.webmvc.servlet.mvc.asis;
-
-import com.interface21.webmvc.servlet.mvc.HandlerMapping;
-import jakarta.servlet.http.HttpServletRequest;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class ControllerHandlerMapping implements HandlerMapping {
-
-    private static final Logger log = LoggerFactory.getLogger(ControllerHandlerMapping.class);
-
-    private static final Map<String, Controller> controllers = new HashMap<>();
-
-    @Override
-    public void initialize() {
-        log.info("Initialized Handler Mapping!");
-        controllers.keySet()
-                .forEach(path -> log.info("Path : {}, Controller : {}", path, controllers.get(path).getClass()));
-    }
-
-    @Override
-    public Optional<Object> getHandler(final HttpServletRequest request) {
-        final var requestURI = request.getRequestURI();
-        log.debug("Request Mapping Uri : {}", requestURI);
-        return Optional.ofNullable(controllers.get(requestURI));
-    }
-
-    public void addController(final String path, final Controller controller) {
-        controllers.put(path, controller);
-    }
-}
